fix adjustment value handling and comparisions, since it uses doubles in
Tue Jan 29 15:19:20 2002  Tim Janik  <timj@gtk.org>

        * beast-gtk/bstdial.[hc]: fix adjustment value handling and
        comparisions, since it uses doubles in gtk+-2.0.

        * beast-gtk/bstmain.c: add --no-plugins option to speed up
        debugging.

svn path=/trunk/; revision=334

+++ beast-gtk/bstdial.h
@@ -60,13 +60,13 @@ struct _BstDial
   guint timer;
   
   /* Current angle of the pointer */
-  gfloat angle;
+  gdouble angle;
   
   /* Old values from GtkAdjustment, stored so we know when something changed */
-  gfloat old_value;
-  gfloat old_lower;
-  gfloat old_upper;
-  gfloat old_page_size;
+  gdouble old_value;
+  gdouble old_lower;
+  gdouble old_upper;
+  gdouble old_page_size;
   
   /* The adjustment object that stores the data for this dial */
   GtkObject *adjustment;
